@@ -17,8 +17,11 @@
   },
   "dependencies": {
     "babel-preset-stage-1": "^6.1.18",
-    "react": "15.4.0",
+    "ramda": "^0.23.0",
+    "react": "^15.5.4",
     "react-d3": "^0.4.0",
-    "react-dom": "^15.5.4"
+    "react-dom": "^15.5.4",
+    "react-router": "^4.1.1",
+    "react-router-dom": "^4.1.1"
   }
 }
