@@ -1,8 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { BrowserRouter as Router, Route } from 'react-router-dom';
+
 import App from './components/App';
 
 ReactDOM.render(
-  <App />,
+  <Router className='route-root'>
+    <Route path='/' component={App} />
+  </Router>,
   document.getElementById('root')
 );
